auth-mvp-fastapi/pyproject.toml
[tool.poetry]
name = "auth-mvp-fastapi"
version = "0.1.0"
description = "MVP Auth: JWT access + refresh rotation"
authors = ["you <you@example.com>"]
readme = "README.md"
packages = [{include="app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.4.0"
SQLAlchemy = "^2.0.34"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
bcrypt = "^4.2.0"
PyJWT = "^2.9.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
httpx = "^0.27.2"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
